+++ stl/lib/src/stl_base.dart
@@ -2,3 +2,4 @@
 class Awesome {
   bool get isAwesome => true;
 }
+//reverse for future implementataion.